DEV: pytest for json comparison
Former-commit-id: 807e3fa127cdd344ad7fa5ca1252ec8d82a46a77

+++ test/test_inference_cpp.cpp
@@ -23,10 +23,10 @@ struct SingleInferenceResult {
 
     float conf;
 
-    float left;
-    float top;
-    float width;
-    float height;
+    int left;
+    int top;
+    int width;
+    int height;
 
     float normalized_left;
     float normalized_top;
